+++ include/boost/spirit/home/qi/detail/expectation_failure.hpp
@@ -1,0 +1,31 @@
+/*=============================================================================
+Copyright (c) 2001-2011 Joel de Guzman
+Copyright (c) 2016      Frank Hein, maxence business consulting gmbh
+
+Distributed under the Boost Software License, Version 1.0. (See accompanying
+file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+=============================================================================*/
+#if !defined(SPIRIT_EXPECTATION_FAILURE_JULY_19_2016)
+#define SPIRIT_EXPECTATION_FAILURE_JULY_19_2016
+
+#if defined(_MSC_VER)
+#pragma once
+#endif
+
+namespace boost { namespace spirit { namespace qi {
+    template <typename Iterator>
+    struct expectation_failure : std::runtime_error
+    {
+        expectation_failure(Iterator first_, Iterator last_, info const& what)
+            : std::runtime_error("boost::spirit::qi::expectation_failure")
+            , first(first_), last(last_), what_(what)
+        {}
+        ~expectation_failure() throw() {}
+
+        Iterator first;
+        Iterator last;
+        info what_;
+    };
+}}}
+
+#endif
