+++ test/qi/expectd.cpp
@@ -20,7 +20,6 @@
 #include <boost/spirit/include/phoenix_core.hpp>
 #include <boost/spirit/include/phoenix_operator.hpp>
 #include <boost/spirit/include/phoenix_statement.hpp>
-#include <mxc/qitoo/qitoo.hpp>
 
 #include <string>
 #include <iostream>
